requirements.txt
# Core - EEG
brainflow>=5.0.0
mne>=1.6.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
pyriemann>=0.9

# Deep Learning (optional, for EEGNet)
torch>=2.0.0
braindecode>=1.3.0

# Nonlinear/Chaos Features
antropy>=0.1.6

# Cursor Control
pyautogui>=0.9.54

# Visualization
pyqtgraph>=0.13.0
PyQt5>=5.15.0
matplotlib>=3.7.0

# Training Paradigm
pygame>=2.5.0

# Configuration
pyyaml>=6.0

# Utilities
joblib>=1.3.0

# Testing
pytest>=7.0.0
